# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Edna Bar Book Printing API
  description: |
    A comprehensive RESTful API for managing book printing operations.
    
    This API provides complete functionality for:
    - Managing books, authors, and publishers
    - Handling printing companies (private and contracted)
    - Creating and tracking printing contracts
    - Scheduling and monitoring printing jobs
    - Real-time delivery tracking
    
    Built with Go following clean architecture principles.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/edna-bar/api
    email: support@edna-bar.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://edna-bar.com/terms
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.edna-bar.com/api
    description: Production server
...
paths:
...
  /printing-jobs:
    get:
      tags:
        - printing-jobs
      summary: List all printing jobs
      description: Retrieve a list of all printing jobs
      parameters:
        - name: status
          in: query
          description: Filter by job status
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, overdue]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrintingJob'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - printing-jobs
      summary: Create a new printing job
      description: Schedule a new printing job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrintingJobRequest'
      responses:
        '201':
          description: Printing job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrintingJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /printing-jobs/overdue:
    get:
      tags:
        - printing-jobs
      summary: Get overdue printing jobs
      description: Retrieve all printing jobs that are past their delivery date
      responses:
        '200':
          description: Overdue jobs found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrintingJob'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    PrintingJob:
      type: object
      properties:
        isbn:
          type: string
          description: Book ISBN
          example: "978-0307474728"
        grafica_id:
          type: integer
          description: Printing company ID
          example: 1
        nto_copias:
          type: integer
          description: Number of copies to print
          example: 5000
        data_entrega:
          type: string
          format: date
          description: Expected delivery date
          example: "2024-02-15"
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
          description: Job status (computed)
          example: "pending"
        days_until_delivery:
          type: integer
          description: Days until delivery (computed)
          example: 30
      required:
        - isbn
        - grafica_id
        - nto_copias
        - data_entrega
    CreatePrintingJobRequest:
      type: object
      properties:
        isbn:
          type: string
          description: Book ISBN
          pattern: '^(97[89])?[\d\-]{10,17}$'
          example: "978-0307474728"
        grafica_id:
          type: integer
          description: Printing company ID
          minimum: 1
          example: 1
        nto_copias:
          type: integer
          description: Number of copies to print
          minimum: 1
          example: 5000
        data_entrega:
          type: string
          format: date
          description: Expected delivery date
          example: "2024-02-15"
      required:
        - isbn
        - grafica_id
        - nto_copias
        - data_entrega
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        details:
          type: string
          description: Additional error details
          example: "Book with ISBN 978-0123456789 was not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            code: "BAD_REQUEST"
            details: "ISBN must be a valid format"
            timestamp: "2024-01-15T10:30:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            details: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Global security (can be overridden per endpoint)
# security:
#   - bearerAuth: []